Create college-draft-ranking-linreg-model.py
This code runs a linear regression on draft data from the previous 5 MLB drafts, examining a plethora of different data features for players. The idea is to examine many different statistics/attributes relevant to college baseball, select features for a model, and then ultimately create the model. The goal of the model is to use draft data to rank players still in college baseball. An analogy can be made between baseball and financial markets, where the draft is the "market" which determines a "market value" (i.e. overall selection number) based on a whole bunch of available information.